clear network code for YMUserAccount

+++ Classes/YMUserAccount.m
@@ -12,17 +12,25 @@
 #import "NSString-SQLiteColumnName.h"
 
 
+
 @implementation YMUserAccount
 
 @synthesize activeNetworkPK, username, password, 
             wrapToken, wrapSecret, loggedIn;
 
+- (void) clearNetworks
+{
+  if (![[SQLiteInstanceManager sharedManager] 
+        tableExists:@"y_m_network"]) return;
+  
+  NSString *q = [NSString stringWithFormat:@"DELETE FROM %@ WHERE %@=%i",
+                 [YMNetwork tableName], [@"userAccountPK" stringAsSQLColumnName], self.pk];
+  [[SQLiteInstanceManager sharedManager] executeUpdateSQL:q];
+}
+
 - (void) deleteObjectCascade:(BOOL)cascade
 {
-  NSString *q = [NSString stringWithFormat:@"DELETE FROM %@ WHERE %@=%i",
-   [YMNetwork tableName], [@"userAccountPK" stringAsSQLColumnName], self.pk];
-  [[SQLiteInstanceManager sharedManager] executeUpdateSQL:q];
-  
+  [self clearNetworks];
   [super deleteObjectCascade:cascade];
 }
